refactor: consolidate JS methods into a single util file
- Merged three separate JS files, each containing a single method, into one util file.
- Moved the new util file into the util folder for better organization.
